Added a byte getter to ByteArray and cleaned up some size types

+++ include/R_String.h
@@ -22,7 +22,7 @@ const char* R_String_getString(R_String* self);
 R_String* R_String_appendCString(R_String* self, const char* string);
 R_String* R_String_appendString(R_String* self, R_String* string);
 R_String* R_String_setString(R_String* self, const char* string);
-R_String* R_String_setSizedString(R_String* self, const char* string, int stringLength);
+R_String* R_String_setSizedString(R_String* self, const char* string, size_t stringLength);
 
 
 R_String* R_String_appendInt(R_String* self, int value);
